Move Map Loading To Its Own Module
